+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### [v1.4.0]
 
-> 2022-03-18
+> 2022-05-09
+
+-   On download page, if url has param like utm_source, utm_medium, utm_campaign, then when click on download button, user will be redirected to App Store or Play Store along with the utm param
 
 -   Re-add promo page